Convert to pbr
Use pbr for packaging instead of manual setup tools. This allows us to
take advantage of a number of features provided by pbr, such as
automated generation of AUTHORS, HISTORY and MANIFEST files.

This removes two doc files due to reliance on the removed AUTHORS and
HISTORY files.

Finally, the 'full' tox target is removed and the 'pyN' targets are
stubbed out. These are broken by the move to pbr and will be rendered
irrelevant by a follow-up change.

Change-Id: I098979a6033afcff9abfbf60de63825f073dc15b
Partial-Bug: #1535772