Convert to pbr
Use pbr for packaging instead of manual setup tools. This allows us to
take advantage of a number of features provided by pbr, such as
automated generation of AUTHORS, HISTORY and MANIFEST files.

Note that tests will not work with this version due to missing entry
points. These will be fixed in the subsequent commit.

Change-Id: I098979a6033afcff9abfbf60de63825f073dc15b