remove test_oversubscribed_server, not useful
This test should fail with Forbidden exception when trying
to spawn more servers than allowed vCPU quota is. This is
because each vcpu->cpu pinned server requires a dedicated
cpu from the host. If spawning 5 servers with 4 pinned
vCPUs each it would over allowed in quota or available
on the host 16 or 20 pCPU (depending on the setup).
Currently when this test is run in parallel
(tempest/testr concurrency) to other tests it doesn't
hit the quota limits but instead fails with
'no valid host was found' instead of Forbidden exception.
Lowering quota will limit other tests so it's not
an option either.

Change-Id: Ib5fdad6f941ea92864f048938021b16dbb41478c
Related-Bug: 1571723