fix get_placement function to use ps and new cgroup path
Change in path to cgroup that has vcpu placement
information is required for the function to work properly
when run on host VM with Ubuntu 14. This breaks this
function run in a docker container where cgroup paths
are different. get_placement function now uses
ps instead of pgrep due to fact pgrep was picking
up more than qemu/VM process - it was not filtering
'/bin/sh pgrep...' line.

Change-Id: I7772dee3f9cf8680a65681fe642eba6e0724f80e
Closes-Bug: #1573157