Rework 'test_numa_topology'
This is a little cludgy and could make better use of existing test
classes found in Tempest. The main changes are:

- The test class is no longer unnecessarily split into a base class and
  test class
- The test class now inherits from 'base.BaseV2ComputeAdminTest' rather
  than from 'manager.NetworkScenarioTest', allowing us to remove a lot
  of boilerplate
- 'remote_client.RemoteClient' is subclassed to indicate how we're
   retrieving host NUMA topology
- Some variables are renamed to clarify their purpose.

This adds a dependency of the '[validation] run_validation' config
option, which must now be set to 'True' to ensure networks are
provisioned for guest instances that need them.

Change-Id: Idfa521e81aba93c863b8996b6f42645711312759