Support more than an allinone deployment
Previously, it was assumed that instances ran on the same host as the
tempest tests. This patch adds support for instances running on
different compute hosts, thus enabling dpeloyments other than an
allinone. Specifically, this adds config options for the SSH user and
key to use when SSH'ing to the deployment's hosts and rewrites various
functions that exectutes commands on the local host to use SSH to
connect to the instance's hypervisor's IP address.

Change-Id: I07ddf7c969a81ebd7fdfda3e4d62cf8a3ad968f8